% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-space.R
\name{apply_zstack}
\alias{apply_zstack}
\title{Apply slice stacking order}
\usage{
apply_zstack(vbl2D, zstack)
}
\arguments{
\item{vbl2D}{A \link[=vbl_doc_vbl2D]{2D vibble}.}

\item{zstack}{Character scalar. Controls the stacking order when multiple slices
are displayed with an offset-layout.

\itemize{
\item{\code{"asc"}: }{ Slices with lower slice indices are drawn first and may be
occluded by higher-index slices in overlapping offset layouts.}
\item{\code{"desc"}: }{ Slices with higher slice indices are drawn first and may be
occluded by lower-index slices in overlapping offset layouts.}
}}
}
\value{
A reordered \code{vbl2D} object.
}
\description{
Orders a \code{vbl2D} object by slice index to control draw order in stacked plots.
}
