% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-space.R
\name{as_bb2D_df}
\alias{as_bb2D_df}
\title{Convert bb2D to data frame}
\usage{
as_bb2D_df(bb2D)
}
\arguments{
\item{bb2D}{A 2D bounding box as validated by \link{is_bb2D}().}
}
\value{
A tibble with columns \code{cmin}, \code{cmax}, \code{rmin}, and \code{rmax}.
}
\description{
Convert a 2D bounding box (\code{bb2D}) into a one-row tibble with
column and row minima and maxima.
}
