% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble.R
\name{as_vibble}
\alias{as_vibble}
\title{Convert data frame to a vibble}
\usage{
as_vibble(.x, as_of)
}
\arguments{
\item{.x}{A data frame/tibble to convert to a vibble.}

\item{as_of}{The data/version to store the data as of.}
}
\value{
A vibble
}
\description{
This function will convert the input to a vibble
}
