% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-space.R
\name{avg_bb2D}
\alias{avg_bb2D}
\title{Average 2D bounding boxes}
\usage{
avg_bb2D(a, b)
}
\arguments{
\item{a, b}{Two \code{bb2D} objects as validated by \link{is_bb2D}().}
}
\value{
A \code{bb2D} object (list with elements \code{col} and \code{row}) where each limit is the
arithmetic mean of the corresponding limits in \code{bb1} and \code{bb2}.
}
\description{
Compute the element-wise mean of two 2D bounding boxes (\code{bb2D}).
}
\details{
Both inputs must have the same named axes (typically \code{col} and \code{row}). The
lower and upper bounds of each axis are averaged separately.
}
\examples{
vbl <- example_vbl()
vbl2D <- vibble2D(vbl, plane = "axi", slices = 60)
bb_slice <- slice_bb(vbl2D, slice = 60)
bb_screen <- screen_limits(vbl2D, slice = 60)
avg_bb <- avg_bb2D(bb_slice, bb_screen)

}
