% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvibble-colors.R
\name{color_opts_viridis}
\alias{color_opts_viridis}
\title{Viridis options for numeric and discrete color mapping}
\usage{
color_opts_viridis()
}
\value{
Character vector.
}
\description{
Returns a character vector of valid input options for both \code{clrp} and
\code{clrsp} arguments.
}
