% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvibble.R
\name{.ggplane_impl}
\alias{.ggplane_impl}
\title{ggplane implementation}
\usage{
.ggplane_impl(vbl2D, var, ncol, nrow, guide, animate, ...)
}
\arguments{
\item{animate}{Logical. If \code{TRUE}, no facets are created. In \code{ggplane()}, this
is always FALSE. Inside build_ggplot.ggvibble() this can be set to TRUE. Should
only be set to TRUE within \code{render_animation()}.}
}
\description{
ggplane implementation
}
\keyword{internal}
