% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvibble.R
\name{.ggvibble_impl}
\alias{.ggvibble_impl}
\title{ggvibble implementation}
\usage{
.ggvibble_impl(vbl2D, var, ncol, nrow, guide, animate, ...)
}
\arguments{
\item{animate}{Logical. If \code{TRUE}, no facets are created. In \code{ggvibble()}, this
is always FALSE. Inside build_ggplot.ggvibble() this can be set to TRUE. Should
only be set to TRUE within \code{render_animation()}.}
}
\description{
ggvibble implementation
}
\keyword{internal}
