% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utils-internal.R
\name{.reorder_lip}
\alias{.reorder_lip}
\title{Reorder CCS-LIP according to orientation.}
\usage{
.reorder_lip(x, orientation)
}
\arguments{
\item{x}{Named object (typically a list or vector) with names corresponding
to coordinate axes.}

\item{orientation}{Character. Three-letter orientation string (e.g. "RAS").}
}
\value{
The input \code{x}, reordered to match the specified orientation.
}
\description{
Reorders a named object according to a given
orientation string by matching axis directions (R/L, S/I, A/P).
}
