% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vibble2D.R
\name{expand_lim2D}
\alias{expand_lim2D}
\title{Expand limits of a 2D vibble}
\usage{
expand_lim2D(vbl2D, expand)
}
\arguments{
\item{vbl2D}{A \link{vibble2D}.}

\item{expand}{An \link[=is_expand]{expand} value used to expand the limits of the output 2D vibble.}
}
\value{
A modified \code{vbl2D} with updated \code{lim} values for \code{col} and \code{row}.
}
\description{
Expand the 2D spatial limits (\code{lim}) of a \code{vbl2D} object according to an absolute or relative
expand specification. Supports logical, absolute, and relative \link[=vbl_doc_expand]{expand rules}.
}
\examples{
vbl2D <- vibble2D(example_vbl(), plane = "axi", slices = 60)

# Absolute expansion: add 2 units to each side
expand_lim2D(vbl2D, 2L)

# Relative expansion: add 10\% of the range
expand_lim2D(vbl2D, 0.1)

}
