% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-manipulation.R
\name{filter_bb2D}
\alias{filter_bb2D}
\title{Filter a vbl2D by a 2D bounding box.}
\usage{
filter_bb2D(vbl2D, bb2D, null_ok = TRUE)
}
\arguments{
\item{vbl2D}{A \code{vbl2D} object.}

\item{bb2D}{A valid 2D bounding box as checked by \code{is_bb2D()}.}

\item{null_ok}{Logical. Whether \code{NULL} limits in \code{bb2D} are treated as open
bounds.}
}
\value{
A filtered \code{vbl2D} object.
}
\description{
Filters a \code{vbl2D} object to rows whose \code{col} and \code{row} coordinates fall within
a given 2D bounding box.
}
