% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-manipulation.R
\name{filter_bb3D}
\alias{filter_bb3D}
\title{Filter a vbl by a 3D bounding box.}
\usage{
filter_bb3D(vbl, bb3D, null_ok = TRUE)
}
\arguments{
\item{vbl}{A \code{vbl} object.}

\item{bb3D}{A valid 3D bounding box as checked by \code{is_bb3D()}.}

\item{null_ok}{Logical. Whether \code{NULL} limits in \code{bb3D} are treated as open
bounds.}
}
\value{
A filtered \code{vbl} object.
}
\description{
Filters a \code{vbl} object to rows whose spatial coordinates fall within a given
3D bounding box.
}
\seealso{
is_bb3D(),
within_limits()
}
