% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-constants.R
\docType{data}
\name{img_anchors}
\alias{img_anchors}
\title{Default image anchors}
\format{
An object of class \code{list} of length 4.
}
\usage{
img_anchors
}
\description{
Default image anchors
}
\keyword{datasets}
