% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-tests.R
\name{is_bb2D}
\alias{is_bb2D}
\title{Check if object is a 2D bounding box}
\usage{
is_bb2D(x)
}
\arguments{
\item{x}{An object to be tested for being a valid 2D bounding box.}
}
\value{
A logical scalar.
}
\description{
Test whether an object encodes a valid 2D bounding box.
Returns \code{TRUE} if the structure and values fulfill the requirements.
}
\details{
A valid 2D bounding box
\itemize{
\item must be a list
\item must be of length 1 or 2
\item must be named, where names correspond to col and/or row
\item must contain only valid \link[=is_limit]{limits}
}
}
\seealso{
\link{filter_bb3D}()
}
