% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-tests.R
\name{is_limit}
\alias{is_limit}
\title{Check validity of numeric limit specification}
\usage{
is_limit(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
A logical scalar.
}
\description{
Validate whether an object represents a proper numeric limit.
}
\details{
A valid limit
\itemize{
\item is numeric,
\item has length two,
\item contains two distinct values,
}
}
\examples{
# Example 1: Valid limit
is_limit(c(10, 20))

# Example 2: Invalid limit (duplicate values)
is_limit(c(10, 10))


}
\seealso{
\link{within_limits}()
}
