% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-tests.R
\name{is_plane}
\alias{is_plane}
\title{Check plane input}
\usage{
is_plane(x)
}
\arguments{
\item{x}{An object to be tested}
}
\value{
A logical scalar.
}
\description{
Test whether input for recurring argument \code{plane}
is valid.
}
\details{
Input is valid if it is a character scalar and one of
\emph{'sag', 'axi' or 'cor'}.
}
