% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-tests.R
\name{is_slice}
\alias{is_slice}
\alias{is_slice_set}
\alias{is_slice_seq}
\title{Check slice indices and slice sequences.}
\usage{
is_slice(x)

is_slice_set(x)

is_slice_seq(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical scalar.
}
\description{
Utilities to test whether objects represent valid slice indices or sequences
of slice indices.
}
\details{
These helpers distinguish between:

\itemize{
\item{Single slice: } A numeric scalar that represents a non-zero integer slice index.
\item{Slice set: } A numeric vector consisting only of integer slice
indices.
\item{Slice seq: } A slice set that forms a contiguous, gap-free sequence spanning
from its minimum to its maximum value
}
}
\note{
Slices must not be strictly integers but must be unambiguously interpretable
as integers.

\preformatted{

is_slice(as.integer(1))
is_slice(as.double(1))
is_slice(as.double(1.1))       # FALSE (decimal)

is_slice_set(c(10, 11, 12))
is_slice_set(c(10, 11, 12.0))
is_slice_set(c(10, 11, 12.2))  # FALSE (last value is decimal)

is_slice_seq(c(5, 6, 7, 8))
is_slice_seq(c(5.0, 6.0, 7.0))
is_slice_seq(c(5, 7, 8))       # FALSE (missing slice)

}
}
