% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tibble_structure_equal.R
\name{is_tibble_structure_equal}
\alias{is_tibble_structure_equal}
\title{Determine if Two Data Frame Structures are Equal}
\usage{
is_tibble_structure_equal(x, y, ignore_col_order = TRUE)
}
\arguments{
\item{ignore_col_order}{}
}
\description{
Determine if Two Data Frame Structures are Equal
}
\details{
This function was extracted from dplyr (internal function), called
\code{dplyr:::is_compatible_data_frame()}.
}
