% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-manipulation.R
\name{join_vibbles}
\alias{join_vibbles}
\title{Join vibbles}
\usage{
join_vibbles(a, b, .rfn = NULL, join = "full", ...)
}
\arguments{
\item{a}{A vibble data.frame.}

\item{b}{A vibble data.frame or a character path to a NIfTI file or folder.}

\item{.rfn}{Optional renaming function or formula applied to non-coordinate
variables of \code{b}.}

\item{join}{Type of join to perform: \code{"full"}, \code{"inner"}, \code{"left"}, \code{"right"}.}

\item{...}{Additional arguments passed to \code{make_vibble()} when \code{b} is a
character path.}
}
\value{
A vibble.
}
\description{
Join two vibble objects by their voxel coordinates, optionally renaming
variables in the second input and creating it from file if needed.
}
\details{
The function checks that both inputs share the same spatial space, coordinate
limits, and coordinate mapping. If \code{b} is a character path, it is converted
into a vibble via \code{make_vibble()}. An optional renaming function \code{.rfn} is
applied to all non-coordinate variables of \code{b} before the join, if specified.
Variable name collisions are not allowed. The final merge is performed using a full, inner,
left, or right join on \code{ccs_labels} where \code{x} of the input for \verb{<type>_join()}
is the input for \code{a} and \code{y} of the join corresponds to resulting vibble from
input for \code{b}.
}
\examples{
\dontrun{
join_vibbles(a = vibble_a, b = vibble_b)
join_vibbles(a, "path/to/folder", space = "orig")
join_vibbles(a, b, .rfn = ~ paste0("b_", .x))
}

}
