% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvibble-layers.R
\name{layer_crop}
\alias{layer_crop}
\title{Crop a ggplane plot to a filtered voxel extent}
\usage{
layer_crop(.cond, expand = FALSE, ...)
}
\arguments{
\item{.cond}{A logical filter expression evaluated on \code{vbl2D} that determines
the voxels used to define the cropped plot extent.}

\item{expand}{Numeric. Expansion factor passed to \code{expand_bb2D()} to grow the
cropped plot extent.}
}
\value{
A \code{ggvibble_layer} object containing the supplied function. When
added to a \code{ggvibble}, the function is executed and its returned layers are
inserted into the plot.
}
\description{
Crops the plot extent to the bounding box of voxels selected by \code{.cond}.
}
\details{
The layer filters \code{vbl2D} via \code{.cond} and then sets plot limits using
\code{ggplot2::coord_equal()} based on the resulting voxel extent.
}
\examples{
vbl <- example_vbl()

ggplane(vbl, var = "t1", plane = "axi", slices = 100) +
  layer_crop(brain, expand = 0.05)

}
