% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvibble-layers.R
\name{layer_lim}
\alias{layer_lim}
\alias{.layer_lim_impl}
\title{Add the plotting limits}
\usage{
layer_lim(color, fill = NA, slices = NULL, ...)

.layer_lim_impl(vbl2D, color, fill, ...)
}
\arguments{
\item{color}{Color of the rectangle outline.}

\item{fill}{Fill color for the rectangles. Defaults to \code{NA}.}

\item{slices}{Optional numeric vector of slice indices to include.
If \code{NULL}, all slices in the underlying \code{vbl2D} are used.}

\item{...}{Additional arguments passed to \code{ggplot2::geom_rect()}.}

\item{vbl2D}{A \link{vibble2D}.}
}
\value{
A \code{ggvibble_layer} object containing the supplied function. When
added to a \code{ggvibble}, the function is executed and its returned layers are
inserted into the plot.
}
\description{
Add a layer that draws rectangles for the current 2D limits (\code{lim}) of each slice in a \code{vbl2D}.
}
\details{
For each slice, \code{lim_slice()} returns a list with \code{col} and \code{row} limits.
These are converted into \code{cmin}, \code{cmax}, \code{rmin}, and \code{rmax} and passed to \code{ggplot2::geom_rect()}.
This is useful for visualising the effective plotting region and checking limit or offset settings.
}
\keyword{internal}
