% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvibble-layers.R
\name{layer_mask}
\alias{layer_mask}
\title{Add a masking layer}
\usage{
layer_mask(color, .cond = NULL, .by = NULL, opacity = 0.25, name = TRUE, ...)
}
\arguments{
\item{color}{Color used for the mask fill.}

\item{.cond}{A logical filter expression that determines the specific
voxels for which this layer is drawn. The expression is evaluated via
\link[rlang:args_data_masking]{data-masking} semantics with the 2D vibble passed
into this layer by \code{ggplane}(). See \link[=vbl_doc_cond]{Details}.}

\item{.by}{A \link[dplyr:dplyr_tidy_select]{tidy-selection} of columns to
group by before applying the filtering logic of \code{.cond}.}

\item{opacity}{Controls voxel transparency. Accepts constants, ranges, or
data-masked expressions. See \link[=vbl_doc_opacity]{Details}.}

\item{name}{Logical or character scalar. If character, the name with which
the color is associated with in the color legend. If logical, \code{FALSE} prevents
appearance in the legend and \code{TRUE} falls back to the naming default of the
function.}
}
\value{
A \code{ggvibble_layer} object containing the supplied function. When
added to a \code{ggvibble}, the function is executed and its returned layers are
inserted into the plot.
}
\description{
Overlay a logical mask on a \link{ggplane}() plot by filling
voxels.
}
\details{
The condition of \code{.cond} determines for which voxels are included in the mask.
If no condition is provided (\code{.cond = NULL}) this layer masks every voxel in
every slice of the 2D vibble passed to this layer by \code{ggplane()}.
}
