% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvibble-layers.R
\name{layer_outline}
\alias{layer_outline}
\title{Add an outline layer}
\usage{
layer_outline(
  color,
  alpha = 0.9,
  linetype = "solid",
  linewidth = 1,
  use_dbscan = TRUE,
  concavity = 2.5,
  clip_overlap = TRUE,
  .cond = NULL,
  .by = NULL,
  ...
)
}
\arguments{
\item{color}{Character. Controls the color used for the lines.}

\item{alpha}{Numeric. Controls the transparency of the lines.}

\item{linetype}{Character. Line pattern for outlines. Supported values:
\itemize{
\item {Named types:}{ c("solid", "dashed", "dotted", "dotdash", "longdash", "twodash") }
\item {Numeric codes:}{ c("0", "1", "2", "3", "4", "5", "6") }
\item {Custom:}{ c("12", "1F", "F0F0") as examples of hexadecimal dash patterns}
}}

\item{linewidth}{Numeric. Controlls the thickness of drawn lines.}

\item{use_dbscan}{Logical. If \code{TRUE}, for every slice \link[dbscan:dbscan]{DBSCAN}
is used to identify spatial islands of voxels matching the condition which
are then outlined separately.}

\item{concavity}{Numeric. A relative measure of concavity. Passed to \link[concaveman:concaveman]{concaveman()}.
1 results in a relatively detailed shape, Infinity results in a convex hull. You can use values lower than 1,
but they can produce pretty crazy shapes.}

\item{clip_overlap}{Logical. Applies only to offset layouts. If TRUE,
content from each slice is clipped so that it cannot extend into adjacent
slice regions.}

\item{.cond}{A logical filter expression that determines the specific
voxels for which this layer is drawn. The expression is evaluated via
\link[rlang:args_data_masking]{data-masking} semantics with the 2D vibble passed
into this layer by \code{ggplane}(). See \link[=vbl_doc_cond]{Details}.}

\item{.by}{A \link[dplyr:dplyr_tidy_select]{tidy-selection} of columns to
group by before applying the filtering logic of \code{.cond}.}
}
\value{
A \code{ggvibble_layer} object containing the supplied function. When
added to a \code{ggvibble}, the function is executed and its returned layers are
inserted into the plot.
}
\description{
Overlay a \link{ggplane}() plot by outlining voxels
that match a certain condition. Outlines are computed slice-wise.
}
\details{
The condition of \code{.cond} determines which voxels are outlined. If no condition
is provided (\code{.cond = NULL}) this layer outlines all voxels in every slice of the
2D vibble passed to this layer by \code{ggplane()}.
}
