% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvibble-layers.R
\name{layer_slice_numbers}
\alias{layer_slice_numbers}
\title{Add slice numbers}
\usage{
layer_slice_numbers(
  anchor = waiver(),
  ref_bb = "screen",
  wrap = "{slice}",
  angle = 0,
  alpha = 0.8,
  color = "white",
  size = 3.5,
  ...
)
}
\arguments{
\item{anchor}{\link[=vbl_doc_img_anchor]{Image anchor} specification for
positioning the slice labels. If, character one of \emph{c('center', 'top', 'bottom', 'left', 'right')}.
Absolute anchors are not allowed in offset-layouts.}

\item{ref_bb}{Character scalar indicating which \link[=vbl_doc_ref_bb]{2D reference bounding box}
to use when anchoring slice numbers via relative or character image anchors:
\itemize{
\item \code{"data"}: uses global bounding boxes from \link{data_bb}().
\item \code{"slice"}: uses per-slice bounding boxes from \link{slice_bb}().
\item \code{"screen"}: uses screen-space bounding boxes from \link{screen_bb}().
}}

\item{wrap}{Optional template for label text. Defaults to \code{"{slice}"}.}

\item{angle}{Rotation angle passed to \code{geom_text()}.}

\item{alpha, color, size}{Passed to \code{geom_text()} to define the label aesthetics.}

\item{...}{Additional arguments passed to \code{geom_text()}.}
}
\value{
A \code{ggvibble_layer} object containing the supplied function. When
added to a \code{ggvibble}, the function is executed and its returned layers are
inserted into the plot.
}
\description{
Add slice numbers as text labels to a \link{ggplane()}-plot.
}
\details{
If \code{align} is used, the selected axis
(\code{col} or \code{row}) is harmonized across all slices.
}
