% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvibble-layouts.R
\name{length.ggvibble_layout}
\alias{length.ggvibble_layout}
\title{Length of a ggvibble layout}
\usage{
\method{length}{ggvibble_layout}(x)
}
\arguments{
\item{x}{A \code{ggvibble_layout}.}
}
\value{
An integer.
}
\description{
S3 method for \code{length()} returning the number of contained plots.
}
