% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{mid}
\alias{mid}
\title{Compute the midpoint of a limit}
\usage{
mid(limit)
}
\arguments{
\item{limit}{A \link[=is_limit]{limit}.}
}
\value{
A numeric scalar giving the midpoint of \code{limit}.
}
\description{
Computes the numeric midpoint of a valid limit object.
}
