% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-space.R
\name{offset_attr}
\alias{offset_attr}
\alias{offset_dir}
\alias{offset_dist}
\title{Access offset-related attributes of a vbl2D object}
\usage{
offset_attr(vbl2D, which)

offset_dir(vbl2D)

offset_dist(vbl2D)
}
\arguments{
\item{vbl2D}{A \link{vibble2D}.}

\item{which}{Name of the offset attribute to extract. Valid options are \emph{c('offset_dir', 'offset_dist', 'offset_plane')}.}
}
\value{
The stored offset attribute value, or \code{NULL} if not present.
}
\description{
Retrieve offset metadata attached to a \code{vbl2D} object.
}
\details{
Offsets are applied in \verb{\link{vibble2D}()} when \code{offset_dist > 0}.
The following attributes can then be accessed:
\itemize{
\item \code{offset_dir}: Direction of voxel shift.
\item \code{offset_dist}: Distance of the applied offset.
\item \code{offset_plane}: Plane with respect to which the offset was defined.
}

If the input object is not offset, a warning is issued
but the function still attempts to return the attribute.
}
\seealso{
\link{vibble2D}(), \link{is_offset}()
}
