% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-constants.R
\docType{data}
\name{offset_dir_choices}
\alias{offset_dir_choices}
\title{Offset directions}
\format{
An object of class \code{character} of length 8.
}
\usage{
offset_dir_choices
}
\value{
A character vector with allowed offset directions.
}
\description{
Contains the valid direction choices used for offsetting 2D vibbles
with \code{offset_dir}.
}
\keyword{datasets}
