% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-constants.R
\docType{data}
\name{offset_directions}
\alias{offset_directions}
\title{Offset directions}
\format{
An object of class \code{character} of length 4.
}
\usage{
offset_directions
}
\value{
A character vector with allowed offset directions.
}
\description{
Contains the valid direction choices used for offsetting 2D vibbles
with \code{offset_dir}.
}
\keyword{datasets}
