% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvibble-layouts.R
\name{Ops.ggvibble_layout}
\alias{Ops.ggvibble_layout}
\title{Operators for ggvibble layouts}
\usage{
\method{Ops}{ggvibble_layout}(e1, e2)
}
\arguments{
\item{e1}{A \code{ggvibble_layout}.}

\item{e2}{The right-hand side operand.}
}
\value{
A modified \code{ggvibble_layout}.
}
\description{
S3 method for the \code{Ops} group generic.

Supported operators:
\itemize{
\item \code{+} : add a \code{ggvibble} to the layout, or add \code{ggvibble_layer}(s) to the last plot.
\item \code{/} : combine the layout with another \code{ggvibble} using \code{/}.
\item \code{|} : combine the layout with another \code{ggvibble} using \code{|}.
\item \code{*} : add a \code{ggvibble_layer} (or list of them) to all plots in the layout.
}
}
