% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvibble-layouts.R
\name{plot.ggvibble_layout}
\alias{plot.ggvibble_layout}
\title{Plot a ggvibble layout}
\usage{
\method{plot}{ggvibble_layout}(x, ...)
}
\arguments{
\item{x}{A \code{ggvibble_layout}.}

\item{...}{Unused.}
}
\value{
Invisibly returns the composed plot object.
}
\description{
S3 method for \code{plot()} that evaluates a \code{ggvibble_layout} by converting each
contained \code{ggvibble} to a ggplot (via \code{as_ggplot()}) and combining them using
the stored operators (\code{+}, \code{/}, \code{|}).
}
