% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvibble-layouts.R
\name{print.ggvibble_layout}
\alias{print.ggvibble_layout}
\title{Print a ggvibble layout}
\usage{
\method{print}{ggvibble_layout}(x, ...)
}
\arguments{
\item{x}{A \code{ggvibble_layout}.}

\item{...}{Passed to \code{plot()}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
S3 method for \code{print()} that forwards to \code{plot()}.
}
