% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{vbl_def}
\alias{vbl_def}
\title{Use vibble default}
\usage{
vbl_def()
}
\value{
An internal sentinel object.
}
\description{
Sentinel value indicating that a parameter should use the global vibble
default as defined via vbl_opts().
}
