% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggvibble.R
\name{vbl_doc_expand}
\alias{vbl_doc_expand}
\alias{is_expand}
\alias{is_expand_lgl}
\alias{is_expand_abs}
\alias{is_expand_rel}
\title{Expanding limits}
\usage{
is_expand(x)

is_expand_lgl(x)

is_expand_abs(x)

is_expand_rel(x)
}
\arguments{
\item{x}{Object to test for being a valid expand specification.}
}
\value{
\itemize{
\item \code{is_expand()}: Logical scalar, \code{TRUE} if \code{x} is a valid expand specification.
\item \code{is_expand_lgl()}: Logical scalar, \code{TRUE} if \code{x} is interpreted as a logical expand rule.
\item \code{is_expand_abs()}: Logical scalar, \code{TRUE} if \code{x} is interpreted as an absolute expand rule.
\item \code{is_expand_rel()}: Logical scalar, \code{TRUE} if \code{x} is interpreted as a relative expand rule.
}
}
\description{
Validate and classify \code{expand} specifications used to modify \link[=is_limit]{limits}.
}
\details{
Expand specifications are vectors of length one or two. If length one, and a second value is
required (as in expansions along col/row) the value is recycled.

Valid expand types are.
\itemize{
\item Absolute: \code{is_expand_abs(x)} is \code{TRUE} for \link[=is_abs]{absolute} inputs of length one or two.
Values are interpreted as absolute units added to both sides of the limits.
\item Relative: \code{is_expand_rel(x)} is \code{TRUE} for  \link[=is_rel]{relative} inputs of length one or two.
Values are interpreted as proportions of the current range added to both sides.
\item Logical: \code{is_expand_lgl(x)} is \code{TRUE} for logical inputs of length one or two.
A \code{TRUE} value expands each side by a fixed fraction of the current range. The factor
is taken from \code{vbl_opts("expand.fraction")}.
}

\code{is_expand(x)} returns \code{TRUE} if any of the above tests is satisfied.
}
\examples{
is_expand(TRUE)
is_expand(c(2L, 4L))
is_expand(c(0.1, 0.2))

is_expand_lgl(TRUE)
is_expand_abs(2L)
is_expand_rel(0.05)

}
