% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-doc.R
\name{vbl_doc_opacity}
\alias{vbl_doc_opacity}
\title{Opacity options}
\description{
The \code{opacity} parameter supports constant, ranged, and voxel-wise inputs.
In all cases, the used values are guaranteed to lie within \verb{[0,1]},
where \code{0} corresponds to zero opacity (completely invisible) and \code{1}
corresponds to full opacity (no shine-through).

It is interpreted as follows:

\itemize{

\item \strong{Single numeric value: }
A constant opacity in \verb{[0,1]} applied to all voxels.
No scaling is performed. Example A.

\item \strong{Numeric vector of length 2: }
Interpreted as a lower and upper opacity bound. The numeric variable being plotted
is rescaled into this interval, with the input range given by \link{var_limits}\verb{(vbl2D, var)}.
Restricted to \link{layer_numeric_var()}. Example B.

\item \strong{Data-masked expressions: }
Any expression using column names of \code{vbl2D} are evaluated with
\link{eval_tidy}(), meaning variable names refer to columns rather than
R objects. See \link[rlang:args_data_masking]{data-masking}. Example C.

\item \strong{Invalid input: }
Non-numeric inputs or numeric vectors of unsupported length cause an error.
Example D.

}
}
