% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utils-internal.R
\name{vbl_doc_resolve_def}
\alias{vbl_doc_resolve_def}
\alias{.resolve_label_spacer}
\alias{.resolve_legend_label}
\alias{.resolve_n}
\alias{.resolve_plane}
\alias{.resolve_verbose}
\title{Resolve parameters with vbl_def()}
\usage{
.resolve_label_spacer(label_spacer)

.resolve_legend_label(label, color, default)

.resolve_n(n, opt)

.resolve_plane(plane)

.resolve_verbose(verbose)
}
\description{
Resolving parameters that can have vbl_def() as default, which
draws from vbl_opts(). Functions should check whether the resolving
works and return an informative error message if not.
}
\keyword{internal}
