% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-doc.R
\name{vbl_doc_var_label}
\alias{vbl_doc_var_label}
\title{Dummy documentation for labeled voxel variables}
\arguments{
\item{var}{Character. The name of a label (factor) column.}
}
\description{
Dummy documentation for labeled voxel variables
}
\keyword{internal}
