% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-doc.R
\name{vbl_doc_var_mask}
\alias{vbl_doc_var_mask}
\title{Dummy documentation for logical mask variables}
\arguments{
\item{var}{Character. The name of a mask (logical) column.}
}
\description{
Dummy documentation for logical mask variables
}
\keyword{internal}
