% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-doc.R
\name{vbl_doc_var_numeric}
\alias{vbl_doc_var_numeric}
\title{Dummy documentation for numeric voxel variables}
\arguments{
\item{var}{Character. The name of a numeric column.}
}
\description{
Dummy documentation for numeric voxel variables
}
\keyword{internal}
