% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-doc.R
\docType{class}
\name{vbl_doc_vbl}
\alias{vbl_doc_vbl}
\title{Vibble objects}
\description{
A vibble is a voxel-level tidy-data structure (tibble/data.frame) that represents
one or more 3D volumes registered to the same space.

\itemize{
\item{Rows}{ correspond to voxels, identified by integer coordinates.}
\item{Columns}{ store voxel-wise values (e.g. intensities, masks, labels, scores).}
}

Every column is referred to as variable in tidy-data terms.
}
\section{Variables}{

A vibble knows four types of variables:

\itemize{
\item{Spatial:}{ Variables with \emph{fixed naming} x, y and z give the unique position of each voxel in the 3D space.
They serve as both, for spatial localization and as a unique identifier.}
\item{Labels:}{ Variables of class \link{factor} storing categorical classifications or ordinal scores.}
\item{Mask:}{ Variables of class \link{logical} storing binary information.}
\item{Numeric:}{ Variables of class \link{numeric} storing information that can be interpreted on a continuous numeric scale.}
}

Labels, mask and numeric variables are referred to as \link[=vars_data]{data variables}. Spatial variables
x, y, and z are referred to as \link[=vars_spatial]{spatial variables}.

For a vibble to be valid, the following requirements must hold.
}

\section{Required class}{

Vibbles extent the \link[tibble:tibble]{tbl} class with \code{c("vbl", <underlying class>)}.
}

\section{Required columns}{

A vibble must contain at least the following columns:
\itemize{
\item{ \code{x}, \code{y}, \code{z}: integer voxel coordinates in a
Cartesian grid. This coordinate triplet must uniquely identify each row.}
\item{ One additional column with voxel values (labels, mask or numeric) }.
}
}

\section{Required spatial mapping}{

The spatial information hold by the \code{x}, \code{y} and \code{z} variables is fixed and
maps to the anatomical axes as follows:

\itemize{
\item{x:}{ Gives the integer position of a voxel along the right-left-axis. Increasing values move
the voxel towards the \strong{left}.}
\item{y:}{ Gives the integer position of a voxel along the superior-inferior-axis. Increasing values move
the voxel towards \strong{inferior}.}
\item{z:}{ Gives the integer position of a voxel along the anterior-posterior-axis. Increasing values move
the voxel towards \strong{posterior}.}
}

Accordingly, XYZ in a Cartesian coordinate system map to LIP in voxel orientation code.
}

\section{Required attributes}{

A vibble must contain the following attributes:

\itemize{
\item \code{ccs_limits}: A list of length three containing the valid
\link[=is_limit]{limits} for \code{x}, \code{y}, and \code{z}. These define the spatial
extent of the Cartesian coordinate system (the “space”) in which the
volume data resides. Vibbles can only be \link[=join_vibbles]{joined} when their
spatial limits are identical.

\item \code{ccs_steps}: A list of length three containing the physical step
size (in millimeters) corresponding to an increase of 1 unit along each of
the three axes.
}
}

\section{Behaviour}{

A vibble behaves like a tibble, except for one thing - Spatial variables x, y and z can not be renamed, removed or manipulated.
}

