% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-doc.R
\name{vbl_doc_vbl_layer}
\alias{vbl_doc_vbl_layer}
\title{Dummy documentation for vibble layers}
\arguments{
\item{cond}{Optional logical filter expression that determines the specific
voxels on which this layer is drawn. This allows fine-grained, voxel-level
filtering beyond whole-slice selection. The expression is evaluated with
\link[rlang:args_data_masking]{tidyverse's data-masking} semantics
on the \link{vibble2D}() passed into the layer by the \link{ggplane}() call.
See \link[=ggvibble_doc_cond]{Details} for more information and examples.

If \code{NULL}, the layer is drawn for all voxels included in the corresponding
\link{ggplane}() call.}

\item{slices}{Optional numeric vector of slice indices on which this layer
should be drawn. This restricts the layer at the slice level, in contrast to
\code{cond}, which restricts the layer at the voxel level.

If \code{NULL}, the layer is applied to all slices included in the corresponding
\link{ggplane}() call.}
}
\value{
A \code{ggvibble_layer} object containing the supplied function. When
added to a \code{ggvibble}, the function is executed and its returned layers are
inserted into the plot.
}
\description{
Dummy documentation for vibble layers
}
\keyword{internal}
