% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{versions}
\alias{versions}
\title{Return list of versions}
\usage{
versions(v)
}
\arguments{
\item{v}{A vibble}
}
\value{
A sorted list of versions in the vibble.
}
\description{
Return list of versions
}
\examples{
if (FALSE) {
   versions(vibble::vibble(iris))
}
}
