% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_snapshot.R
\name{add_snapshot}
\alias{add_snapshot}
\title{Add a snapshot}
\usage{
add_snapshot(v, snapshot, as_of = lubridate::now())
}
\arguments{
\item{v}{A vibble}

\item{snapshot}{A new table with the same structure as vibble (without version data).}

\item{as_of}{A date/version}
}
\value{
An updated vibble with newtbl merged in.
}
\description{
Given a data frame, update the vibble by adding a snapshot at the
current version point.
}
\details{
The algorithm involves adding a new version of all rows, then combining identical rows
so that versions are listed together (in a vlist).
}
\examples{
\dontrun{
vibble::add_snapshot(vibble::vibble(iris, as_of="2022-01-01"), iris[1:10,], as_of="2022-01-02")
}
}
