% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_snapshot.R
\name{add_snapshot_with_different_structure}
\alias{add_snapshot_with_different_structure}
\title{Add snapshot with different structure to vibble}
\usage{
add_snapshot_with_different_structure(v, snapshot, as_of)
}
\arguments{
\item{as_of}{}
}
\description{
Add snapshot with different structure to vibble
}
\details{
This function assumes that the vibble and the snapshot have different structures. As such,
there is no way to reuse data from prior time points; everything is invalidated for \code{as_of} point.

The function does have to conform the structure, which conveniently \code{\link[dplyr:bind]{dplyr::bind_rows()}} does without
a problem.
}
