% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_snapshot.R
\name{add_snapshot_with_same_structure}
\alias{add_snapshot_with_same_structure}
\title{Add snapshot with same structure to vibble}
\usage{
add_snapshot_with_same_structure(v, snapshot, as_of)
}
\arguments{
\item{as_of}{}
}
\description{
Add snapshot with same structure to vibble
}
