% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvibble.R
\name{as_ggplot}
\alias{as_ggplot}
\title{Convert a ggvibble to a ggplot object}
\usage{
as_ggplot(p, ...)
}
\arguments{
\item{x}{A \code{ggvibble} object.}
}
\value{
A ggplot object that can be further modified using ggplot2 or composed using
patchwork.
}
\description{
Materializes a \code{ggvibble} object into a ggplot object.
}
\details{
\code{as_ggplot()} terminates the ggvibble layer pipeline and converts it into a
concrete ggplot representation. After conversion, ggvibble layers can no
longer be added.
}
