% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_of.R
\name{as_of}
\alias{as_of}
\title{Extract a data frame from a vibble as of a certain version}
\usage{
as_of(v, as_of = NULL)
}
\arguments{
\item{v}{A vibble}

\item{as_of}{A date to construct the snapshot from (assumed YYYY-MM-DD).}
}
\value{
A data frame with contents representing a snapshot as of a specific time.
}
\description{
Given a table with information over time,
retrieve data as of a specific time point.
}
\details{
The design pattern is documented in many places. For one description, see
https://docs.microsoft.com/en-us/sql/relational-databases/tables/temporal-tables?view=sql-server-2017
}
\examples{
\dontrun{
vibble::as_of(vibble::vibble(iris, as_of="20220101"),"20220101")
}
}
