% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvibble.R
\name{ggplane}
\alias{ggplane}
\title{Build a 2D ggplot representation of voxel data}
\usage{
ggplane(
  vbl,
  var,
  plane,
  slices,
  lim = NULL,
  expand = 0.1,
  offset_dist = 0,
  offset_dir = "right",
  flip = FALSE,
  clrp = vbl_opts("clrp"),
  clrsp = vbl_opts("clrsp"),
  interpolate = vbl_opts("interpolate"),
  layout = NULL,
  .cond = NULL,
  ...
)
}
\arguments{
\item{vbl}{A \link{vibble}.}

\item{var}{Character scalar naming the variable to plot.}

\item{plane}{The anatomical orientation. Valid options are \emph{c('sag', 'axi', 'cor')}.}

\item{slices}{Optional. Integer vector of slice positions in the selected \code{plane}
to keep. If \code{NULL}, all slices are retained.}

\item{lim}{Defines the 2D spatial limits of the resulting \code{vbl2D}. By default, the
\link[=ccs_limits]{coordinate-space} limits of the input vibble are and mapped to
col, row and slice depending on \code{plane}. User can adjust that by supplying:

\itemize{
\item  a valid \link[=is_limit]{limit}: a numeric vector of length two which is recycled for col and row.
\item a list with \code{col} and \code{row} elements, each being a \link[=is_limit]{limit} for the respective axis.
}

When the supplied \link[=is_bb2D]{2D bounding-box} is smaller than the slice extents,
only voxels inside this region are retained, giving \code{lim} a filtering effect.}

\item{expand}{An \link[=is_expand]{expand} specification applied to the col/row limits \strong{after}
the interpretation and potential filtering of \code{lim}. This enlarges the slice-wise bounding box by adding a
margin around its current range.}

\item{offset_dist}{If not 0, shifts the slices in the 2D layout.
Interpreted as:
\itemize{
\item \strong{absolute} when supplied as an integer (e.g. \code{2L}) or marked via \code{as_abs()}.
\item \strong{relative} when supplied as a double (e.g. \code{0.1}) or marked via \code{as_rel()},
in which case the shift is computed as a proportion of the full axis range.
}}

\item{offset_dir}{Direction in which to shift voxels when \code{offset_dist} is not 0.
Valid options are \emph{c('left', 'right', 'top', 'bottom')}.}

\item{flip}{Logical. If \code{TRUE}, the z-order of the slices is flipped. Only relevent for offset-layouts.}

\item{clrp}{Character scalar specifying the categorical palette.
Must be one of \code{\link{clrp_opts_vec}()}. Defaults to \code{vbl_opts("clrp")}.}

\item{clrsp}{Character scalar specifying the numeric color palette used in \link{scale_fill_numeric}()`.}

\item{interpolate}{Logical scalar indicating whether to interpolate raster tiles.}

\item{layout}{Optional numeric vector of length two specifying \code{c(ncol, nrow)} for \code{facet_wrap()}
when multiple slices are shown and \code{offset_dist} ist 0.}

\item{.cond}{Optional. An additional logical filter expression that determines the specific
voxels plotted with \code{ggplane()} and passed to added \link[vbl_doc_vbl_layer]{layers}. \if{html}{\out{<br>}}

The expression is evaluated via \link[rlang:args_data_masking]{data-masking} semantics.}

\item{...}{Additional arguments forwarded to the scale functions controlling the fill aesthetic.}
}
\value{
A \link{ggvibble} object.
}
\description{
Build a 2D visualization of vocel data by selecting slices in a
specified anatomical plane and plotting a numeric variable as a raster image.
}
\details{
\code{ggplane()} converts \code{vbl} to a \code{vbl2D} object via \link{vibble2D}().
It then stores plotting arguments and returns a \code{ggvibble} container. If multiple slices are present
and the object is not offset, slices are arranged using \link{facet_wrap}() with the requested \code{layout}.

Numeric variables are plotted with \link{scale_fill_numeric}() using \code{clrsp}. Non-numeric variables are
plotted with \link{scale_fill_label}() using \code{clrp}. Note, that in \code{ggplane()}, logical mask variables
are coerced to factors with levels \code{c("TRUE", "FALSE")} before plotting, for binary mask overlays see \link{layer_mask()}.
}
