% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-tests.R
\name{is_limit}
\alias{is_limit}
\title{Check validity of numeric limit specification}
\usage{
is_limit(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
A logical scalar.
}
\description{
Validate whether an object represents a proper numeric limit
consisting of two distinct, non-negative values.
}
\details{
A valid limit
\itemize{
\item is numeric,
\item can be unambiguously interpreted as an integer
\item has length two,
\item contains two distinct values,
\item contains only values greater than or equal to 1.
}
}
\examples{
# Example 1: Valid limit
is_limit(c(10, 20))

# Example 2: Invalid limit (duplicate values)
is_limit(c(10, 10))

# Example 3: Invalid limit (not integer)
is_limit(c(10, 15.4))

}
\seealso{
\link{within_limits}()
}
