% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvibble-layers.R
\name{layer_bb}
\alias{layer_bb}
\alias{layer_bb_data}
\alias{layer_bb_plot}
\alias{layer_bb_screen}
\alias{layer_bb_slice}
\title{Add a data-driven 2D bounding-box}
\usage{
layer_bb(color, .cond, .by = NULL, fill = NA, expand = FALSE, name = TRUE, ...)

layer_bb_data(
  color = alpha("green", 0.7),
  fill = NA,
  slices = NULL,
  name = TRUE,
  ...
)

layer_bb_plot(color = alpha("yellow", 0.7), fill = NA, ...)

layer_bb_screen(color = alpha("blue", 0.7), fill = NA, slices = NULL, ...)

layer_bb_slice(color = alpha("red", 0.7), fill = NA, slices = NULL, ...)
}
\arguments{
\item{color}{Color of the bounding-box outline.}

\item{.cond}{A logical filter expression that determines the specific
voxels for which this layer is drawn. The expression is evaluated via
\link[rlang:args_data_masking]{data-masking} semantics with the 2D vibble passed
into this layer by \code{ggplane}(). See \link[=vbl_doc_cond]{Details}.}

\item{.by}{A \link[dplyr:dplyr_tidy_select]{tidy-selection} of columns to
group by before applying the filtering logic of \code{.cond}.}

\item{fill}{Fill of the bounding-box.}

\item{name}{Logical or character scalar. If character, the name with which
the color is associated with in the color legend. If logical, \code{FALSE} prevents
appearance in the legend and \code{TRUE} falls back to the naming default of the
function.}

\item{...}{Additional arguments passed to \link{geom_rect}().}

\item{slices}{Integer vector. The slices of interest.}
}
\value{
A \code{ggvibble_layer} object containing the supplied function. When
added to a \code{ggvibble}, the function is executed and its returned layers are
inserted into the plot.
}
\description{
Draw a rectangular bounding box around the extent of voxels.
}
\details{
\code{layer_bb()} computes and visualizes a 2D bounding box for voxels based on the
minimal and maximal \code{col} and \code{row} coordinates where \code{.cond} evaluates to \code{TRUE}.
The bounding box can be expanded by \code{buffer}, which adds a proportional
margin to all sides. The graphical rendering is handled by \link{geom_rect}().
}
