% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvibble-layers.R
\name{layer_bb}
\alias{layer_bb}
\title{Add a 2D bounding-box}
\usage{
layer_bb(color, fill = NA, expand = FALSE, .cond = NULL, .by = NULL, ...)
}
\arguments{
\item{color}{Color of the bounding-box outline.}

\item{fill}{Fill of the bounding-box.}

\item{.cond}{Optional. An additional logical filter expression that determines the specific
voxels for which this layer is drawn. The expression is evaluated via
\link[rlang:args_data_masking]{data-masking} semantics with the 2D vibble passed
into this layer by \link{ggplane}(). See \link[=ggvibble_doc_cond]{Details} for more information and examples.}

\item{.by}{Optional. A \link[dplyr:dplyr_tidy_select]{tidy-selection} of columns to
group by before applying the filtering logic of \code{.cond}.}

\item{...}{Additional arguments passed to \link{geom_rect}().}
}
\value{
A \code{ggvibble_layer} object containing the supplied function. When
added to a \code{ggvibble}, the function is executed and its returned layers are
inserted into the plot.
}
\description{
Draw a rectangular bounding box around the extent of voxels
that match a certain condition.
}
\details{
\code{layer_bb()} computes and visualizes a 2D bounding box for voxels based on the
minimal and maximal \code{col} and \code{row} coordinates where \code{.cond} evaluates to \code{TRUE}.
The bounding box can be expanded by \code{buffer}, which adds a proportional
margin to all sides. The graphical rendering is handled by \link{geom_rect}().
}
