% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvibble-layers.R
\name{layer_categorical}
\alias{layer_categorical}
\title{Add a color layer for categorical variables}
\usage{
layer_categorical(
  var,
  clrp = "hue_pal",
  clrp_adjust = NULL,
  opacity = 0.25,
  guide = "drop",
  slices = NULL,
  .cond = NULL,
  .by = "slice",
  ...
)
}
\arguments{
\item{var}{Character. The name of a factor-variable with categorical labels.
If a logical (mask) variable is specified, it is treated as a categorical one
with labels = c('TRUE', 'FALSE').}

\item{clrp}{Character scalar specifying the discrete color palette. \link[=scale_fill_categorical]{Details}.}

\item{clrp_adjust}{Optional named vector of hex colors used to override
colors of specific labels: \verb{c(<label> = <color>, <label2> = <color2>)}}

\item{opacity}{Controls voxel transparency in raster layers. Accepts constants, ranges, or
data-masked expressions. See \link[=vbl_doc_opacity]{Details}.}

\item{slices}{Optional. Numeric vector of \link[=is_slice]{slice numbers} used to
restrict the content of this layer to specific slices. If \code{NULL}, no restriction
is applied and the function considers all slices, as denoted in the \link{ggvibble}()
container.}

\item{.cond}{A logical filter expression that determines the specific
voxels included in the content of this layer. The expression is evaluated via
\link[rlang:args_data_masking]{data-masking} semantics with the 2D vibble passed
into this layer by \code{ggvibble}(). See \link[=vbl_doc_cond]{Details}.}

\item{.by}{A \link[dplyr:dplyr_tidy_select]{tidy-selection} of columns to
group by before applying the filtering logic of \code{.cond}. Defaults to \emph{'slice'}.}

\item{...}{Additional arguments passed to \link{scale_fill_categorical}().}
}
\value{
A \code{ggvibble_layer} object containing the supplied function. When
added to a \code{ggvibble}, the function is executed and its returned layers are
inserted into the plot.
}
\description{
Overlay a categorical label variable on a \code{ggvibble()} plot by
filling voxels with discrete colors.
}
