% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvibble-layers.R
\name{layer_mask}
\alias{layer_mask}
\title{Add a masking layer}
\usage{
layer_mask(
  .cond = NULL,
  color = "red",
  opacity = 0.25,
  label = TRUE,
  slices = NULL,
  .by = "slice",
  ...
)
}
\arguments{
\item{.cond}{A logical filter expression that determines the specific
voxels included in the content of this layer. The expression is evaluated via
\link[rlang:args_data_masking]{data-masking} semantics with the 2D vibble passed
into this layer by \code{ggplane}(). See \link[=vbl_doc_cond]{Details}.}

\item{color}{Character scalar. The \link[=is_color]{color} used for
the mask.}

\item{opacity}{Controls voxel transparency in raster layers. Accepts constants, ranges, or
data-masked expressions. See \link[=vbl_doc_opacity]{Details}.}

\item{label}{Logical or character scalar. Controls the legend label with which
the color of this layer is associated with in the legend.

\itemize{
\item{Character: } Uses the provided value as the label.
\item{\code{TRUE}: } Uses the function’s default to create the label.
\item{\code{FALSE}: } Does not include the color in the legend.
}}

\item{slices}{Optional. Numeric vector of \link[=is_slice]{slice numbers} used to
restrict the content of this layer to specific slices. If \code{NULL}, no restriction
is applied and the function considers all slices, as denoted in the \link{ggplane}()
container.}

\item{.by}{A \link[dplyr:dplyr_tidy_select]{tidy-selection} of columns to
group by before applying the filtering logic of \code{.cond}. Defaults to \emph{'slice'}.}
}
\value{
A \code{ggvibble_layer} object containing the supplied function. When
added to a \code{ggvibble}, the function is executed and its returned layers are
inserted into the plot.
}
\description{
Overlay a logical mask on a \link{ggplane}() plot by filling
voxels.
}
\details{
The condition of \code{.cond} determines which voxels are included in the mask.
If no condition is provided (\code{.cond = NULL}) this layer masks every voxel in
every slice of the 2D vibble passed to it by \code{ggplane()}.
}
