% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvibble-layers.R
\name{layer_outline}
\alias{layer_outline}
\title{Add outlines}
\usage{
layer_outline(
  .cond = NULL,
  color = "gold",
  fill = NA,
  linetype = "solid",
  linewidth = 0.5,
  use_dbscan = TRUE,
  concavity = 1.5,
  label = TRUE,
  slices = NULL,
  .by = "slice",
  ...
)
}
\arguments{
\item{.cond}{A logical filter expression that determines the specific
voxels included in the content of this layer. The expression is evaluated via
\link[rlang:args_data_masking]{data-masking} semantics with the 2D vibble passed
into this layer by \code{ggplane}(). See \link[=vbl_doc_cond]{Details}.}

\item{linetype}{Defines the line pattern. Supported values:
\itemize{
\item {Named types:}{ c("solid", "dashed", "dotted", "dotdash", "longdash", "twodash") }
\item {Numeric codes:}{ c("0", "1", "2", "3", "4", "5", "6") }
\item {Custom:}{ c("12", "1F", "F0F0") as examples of hexadecimal dash patterns}
}}

\item{linewidth}{Numeric scalar. Controlls the thickness of drawn lines.}

\item{use_dbscan}{Logical. If \code{TRUE}, for every slice \link[dbscan:dbscan]{DBSCAN}
is used to identify spatial islands of voxels matching the condition which
are then outlined separately.}

\item{concavity}{Numeric. A relative measure of concavity. Passed to \link[concaveman:concaveman]{concaveman()}.
1 results in a relatively detailed shape, Infinity results in a convex hull. You can use values lower than 1,
but they can produce pretty crazy shapes.}

\item{label}{Logical or character scalar. Controls the legend label with which
the color of this layer is associated with in the legend.

\itemize{
\item{Character: } Uses the provided value as the label.
\item{\code{TRUE}: } Uses the function’s default to create the label.
\item{\code{FALSE}: } Does not include the color in the legend.
}}

\item{slices}{Optional. Numeric vector of \link[=is_slice]{slice numbers} used to
restrict the content of this layer to specific slices. If \code{NULL}, no restriction
is applied and the function considers all slices, as denoted in the \link{ggplane}()
container.}

\item{.by}{A \link[dplyr:dplyr_tidy_select]{tidy-selection} of columns to
group by before applying the filtering logic of \code{.cond}. Defaults to \emph{'slice'}.}

\item{alpha}{Numeric. Controls the transparency of the lines.}
}
\value{
A \code{ggvibble_layer} object containing the supplied function. When
added to a \code{ggvibble}, the function is executed and its returned layers are
inserted into the plot.
}
\description{
Overlay a \link{ggplane}() plot by outlining voxels
that match a certain condition.
}
\details{
The condition of \code{.cond} determines which voxels are outlined. If no condition
is provided (\code{.cond = NULL}) this layer outlines all voxels in every slice of the
2D vibble passed to this layer by \code{ggplane()}.

Outlines are \strong{always} computed slicewise.
}
