% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-misc.R
\name{vars_vibble}
\alias{vars_vibble}
\alias{vars_data}
\alias{vars_label}
\alias{vars_mask}
\alias{vars_numeric}
\alias{vars_spatial}
\alias{vars_type}
\title{Variable selectors for vibble columns}
\usage{
vars_data(vbl)

vars_label(vbl)

vars_mask(vbl)

vars_numeric(vbl)

vars_spatial(vbl)

vars_type(vbl, type)
}
\arguments{
\item{vbl}{A \link{vibble}.}

\item{type}{Character value. The data variable type of interest. Must be one of \emph{c('label', 'mask', 'numeric')}.}
}
\value{
A character vector with column names:

\itemize{
\item \link{vars_data}() names of all non-spatial variables.
\item \link{vars_label}() names of non-spatial variables stored as factors or character labels.
\item \link{vars_mask}() names of non-spatial variables stored as logical masks.
\item \link{vars_numeric}() names of non-spatial variables stored as numeric vectors.
\item \link{vars_spatial}() names of spatial coordinate variables such as \code{x}, \code{y}, \code{z}, \code{col}, \code{row}, and \code{slice}.
\item \link{vars_type}() names of the variables of the requested type.
}

If no variables match the requirements the output is an empty character vector.
}
\description{
Helpers to retrieve column names from a vibble based on their type.
}
