% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-doc.R
\name{vbl_doc}
\alias{vbl_doc}
\title{Dummy documentation for recurring parameters}
\arguments{
\item{lut}{Either a file path to a LUT, character vector of labels or a data.frame with integer indices
and character labels.}

\item{vbl}{A \link{vibble}.}
}
\description{
Dummy documentation for recurring parameters
}
\keyword{internal}
