% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-doc.R
\name{vbl_doc}
\alias{vbl_doc}
\title{Dummy documentation for recurring parameters & sections}
\arguments{
\item{.by}{A \link[dplyr:dplyr_tidy_select]{tidy-selection} of columns to
group by before applying the filtering logic of \code{.cond}.}

\item{bb2D}{A named list that defines a \link[=is_bb2D]{2D bounding box} with \link[=is_limit]{limits} for col and row.}

\item{bb3D}{A named list that defines a \link[=is_bb3D]{3D bounding box} with \link[=is_limit]{limits} for x, y, z.}

\item{clrp}{Character scalar specifying the categorical palette.
Must be one of \code{\link{clrp_opts_vec}()}. Defaults to \code{vbl_opts("clrp")}.}

\item{clrp_adjust}{Optional named vector of named hex colors used to override
colors of specific labels (see \link{color_vector}()).}

\item{clrsp}{Character scalar specifying the numeric color palette used in \link{scale_fill_numeric}().}

\item{concavity}{Numeric. A relative measure of concavity. Passed to \link[concaveman:concaveman]{concaveman()}.
1 results in a relatively detailed shape, Infinity results in a convex hull. You can use values lower than 1,
but they can produce pretty crazy shapes.}

\item{interpolate}{Logical scalar indicating whether to interpolate raster tiles.}

\item{lut}{Either a file path to a LUT, character vector of labels or a data.frame with integer indices
and character labels.}

\item{opacity}{Controls voxel transparency. Accepts constants, ranges, or
data-masked expressions. See \link[=vbl_doc_opacity]{Details}.}

\item{plane}{The anatomical orientation. Valid options are \emph{c('sag', 'axi', 'cor')}.}

\item{rm0}{Logical. If \code{TRUE}, remove voxels with value 0 from the resulting
vibble.}

\item{slice}{Integer value. The slice of interest.}

\item{slices}{Integer vector. The slices of interest.}

\item{vbl}{A \link[=vbl_doc_vbl]{vibble}.}

\item{vbl2D}{A \link[=vbl_doc_vbl2D]{2D vibble}.}

\item{verbose}{Logical. If \code{TRUE}, informative messages are printed in the
console.}
}
\description{
Dummy documentation for recurring parameters & sections
}
\keyword{internal}
