% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-doc.R
\name{vbl_doc}
\alias{vbl_doc}
\title{Dummy documentation for recurring parameters & sections}
\arguments{
\item{.by}{A \link[dplyr:dplyr_tidy_select]{tidy-selection} of columns to
group by before applying the filtering logic of \code{.cond}.}

\item{bb2D}{A named list that defines a \link[=is_bb2D]{2D bounding box} with \link[=is_limit]{limits} for col and row.}

\item{bb3D}{A named list that defines a \link[=is_bb3D]{3D bounding box} with \link[=is_limit]{limits} for x, y, z.}

\item{clrp}{Character scalar specifying the discrete color palette. \link[=scale_fill_categorical]{Details}.}

\item{clrp_adjust}{Optional named vector of hex colors used to override
colors of specific labels: \verb{c(<label> = <color>, <label2> = <color2>)}}

\item{clrsp}{Character scalar specifying the continuous color spectrum if \code{var} refers to a numeric
variable. If not specified, draws from the global default \code{vbl_opts('clrsp')}. \link[=scale_fill_numeric]{Details}.}

\item{concavity}{Numeric. A relative measure of concavity. Passed to \link[concaveman:concaveman]{concaveman()}.
1 results in a relatively detailed shape, Infinity results in a convex hull. You can use values lower than 1,
but they can produce pretty crazy shapes.}

\item{interpolate}{Logical scalar indicating whether to interpolate raster tiles.}

\item{lut}{Either a file path to a LUT, character vector of labels or a data.frame with integer indices
and character labels.}

\item{offset_col, offset_row}{Numeric scalars. Offsets applied per slice along
the \code{col} and \code{row} axes, which can be specified as either \strong{absolute} or \strong{relative} values:

\itemize{
\item{Absolute:}{ Integers, or numeric values wrapped in \code{as_abs()}, are
interpreted as absolute offsets in data coordinates.}
\item{Relative:}{ Numeric values, or values wrapped in \code{as_rel()}, are
interpreted as offsets relative to the \link[=data_bb]{screen bounding box}.}
}}

\item{opacity}{Controls voxel transparency in raster layers. Accepts constants, ranges, or
data-masked expressions. See \link[=vbl_doc_opacity]{Details}.}

\item{plane}{Character scalar. The anatomical orientation, one of \emph{'sag', 'axi' or 'cor'}.
Defaults to the \link[=vbl_doc_defaults]{global default plane}.}

\item{rm0}{Logical. If \code{TRUE}, remove voxels with value 0 from the resulting
vibble.}

\item{slice}{Integer value. The slice of interest.}

\item{slices}{Integer vector. The slices of interest.}

\item{vbl}{A \link[=vbl_doc_vbl]{vibble}.}

\item{vbl2D}{A \link[=vbl_doc_vbl2D]{2D vibble}.}

\item{verbose}{Logical. If \code{TRUE}, informative messages are printed in the
console.}

\item{x}{An object for which a method has been defined.}

\item{zstack}{Character scalar. Controls the stacking order of slices along
the slice axis when multiple slices are displayed with an offset-layout.

\itemize{
\item{\code{"asc"}: }{ Slices with lower slice indices are drawn first and may be
occluded by higher-index slices in overlapping offset layouts.}
\item{\code{"desc"}: }{ Slices with higher slice indices are drawn first and may be
occluded by lower-index slices in overlapping offset layouts.}
}}
}
\description{
Dummy documentation for recurring parameters & sections
}
\keyword{internal}
