% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-doc.R
\name{vbl_doc_layer}
\alias{vbl_doc_layer}
\title{Dummy documentation for vibble layers}
\arguments{
\item{.cond}{A logical filter expression that determines the specific
voxels for which this layer is drawn. The expression is evaluated via
\link[rlang:args_data_masking]{data-masking} semantics with the 2D vibble passed
into this layer by \code{ggplane}(). See \link[=vbl_doc_cond]{Details}.}

\item{linetype}{Character. Line pattern for outlines. Supported values:
\itemize{
\item {Named types:}{ c("solid", "dashed", "dotted", "dotdash", "longdash", "twodash") }
\item {Numeric codes:}{ c("0", "1", "2", "3", "4", "5", "6") }
\item {Custom:}{ c("12", "1F", "F0F0") as examples of hexadecimal dash patterns}
}}

\item{linewidth}{Numeric. Controlls the thickness of drawn lines.}

\item{clip_overlap}{Logical. Applies only to offset layouts. If TRUE,
content from each slice is clipped so that it cannot extend into adjacent
slice regions.}
}
\value{
A \code{ggvibble_layer} object containing the supplied function. When
added to a \code{ggvibble}, the function is executed and its returned layers are
inserted into the plot.
}
\description{
Dummy documentation for vibble layers
}
\keyword{internal}
