% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-doc.R
\name{vbl_doc_layer}
\alias{vbl_doc_layer}
\title{Dummy documentation for vibble layers}
\arguments{
\item{.cond}{Optional. An additional logical filter expression that determines the specific
voxels for which this layer is drawn. The expression is evaluated via
\link[rlang:args_data_masking]{data-masking} semantics with the 2D vibble passed
into this layer by \link{ggplane}(). See \link[=ggvibble_doc_cond]{Details} for more information and examples.}
}
\value{
A \code{ggvibble_layer} object containing the supplied function. When
added to a \code{ggvibble}, the function is executed and its returned layers are
inserted into the plot.
}
\description{
Dummy documentation for vibble layers
}
\keyword{internal}
