% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-doc.R
\name{vbl_doc_layer}
\alias{vbl_doc_layer}
\title{Dummy documentation for vibble layers}
\arguments{
\item{.by}{A \link[dplyr:dplyr_tidy_select]{tidy-selection} of columns to
group by before applying the filtering logic of \code{.cond}. Defaults to \emph{'slice'}.}

\item{.cond}{A logical filter expression that determines the specific
voxels included in the content of this layer. The expression is evaluated via
\link[rlang:args_data_masking]{data-masking} semantics with the 2D vibble passed
into this layer by \code{ggplane}(). See \link[=vbl_doc_cond]{Details}.}

\item{linetype}{Defines the line pattern. Supported values:
\itemize{
\item {Named types:}{ c("solid", "dashed", "dotted", "dotdash", "longdash", "twodash") }
\item {Numeric codes:}{ c("0", "1", "2", "3", "4", "5", "6") }
\item {Custom:}{ c("12", "1F", "F0F0") as examples of hexadecimal dash patterns}
}}

\item{linewidth}{Numeric scalar. Controlls the thickness of drawn lines.}

\item{label}{Logical or character scalar. Controls the legend label with which
the color of this layer is associated with in the legend.

\itemize{
\item{Character: } Uses the provided value as the label.
\item{\code{TRUE}: } Uses the function’s default to create the label.
\item{\code{FALSE}: } Does not include the color in the legend.
}}

\item{offset_clip}{Logical. Applies only to offset layouts. If \code{TRUE},
content from each slice is clipped so that it cannot extend into adjacent
slice regions.}

\item{slices}{Optional. Numeric vector of \link[=is_slice]{slice numbers} used to
restrict the content of this layer to specific slices. If \code{NULL}, no restriction
is applied and the function considers all slices, as denoted in the \link{ggplane}()
container.}
}
\value{
A \code{ggvibble_layer} object containing the supplied function. When
added to a \code{ggvibble}, the function is executed and its returned layers are
inserted into the plot.
}
\description{
Dummy documentation for vibble layers
}
\keyword{internal}
