% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-doc.R
\name{vbl_doc_var_mask}
\alias{vbl_doc_var_mask}
\title{Dummy documentation for logical mask variables}
\arguments{
\item{var}{Character. Must refer to a column in the
\link{vibble} containing \code{TRUE}/\code{FALSE} mask values.}
}
\description{
Dummy documentation for logical mask variables
}
\keyword{internal}
