% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble-doc.R
\name{vbl_doc_var_numeric}
\alias{vbl_doc_var_numeric}
\title{Dummy documentation for numeric voxel variables}
\arguments{
\item{var}{Character. Must refer to a column in the
\link{vibble} containing continuous or integer-valued numeric data.}
}
\description{
Dummy documentation for numeric voxel variables
}
\keyword{internal}
