% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble.R
\name{vibble}
\alias{vibble}
\title{Construct a vibble as of a time point from a data frame}
\usage{
vibble(.x, as_of = lubridate::today())
}
\arguments{
\item{.x}{A tibble/data frame to store}

\item{as_of}{A date or version tag to store as the timepoint.}
}
\value{
A vibble representing \code{.x} at the time point \code{as_of}.
}
\description{
A versioned tibble (vibble) can be created from a data frame at a specific time point.
This function will create this data structure (type \code{tbl_vdf}).
}
\examples{
\dontrun{
vibble(iris, as_of="2022-01-01")
}
}
