% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble2D.R
\name{vibble2D}
\alias{vibble2D}
\title{Create a 2D vibble from 3D voxel data}
\usage{
vibble2D(
  vbl,
  plane,
  slices = NULL,
  lim = NULL,
  expand = FALSE,
  offset_col = 0,
  offset_row = 0,
  .cond = NULL,
  .by = NULL
)
}
\arguments{
\item{vbl}{A \link[=vbl_doc_vbl]{vibble}.}

\item{plane}{The anatomical orientation. Valid options are \emph{c('sag', 'axi', 'cor')}.}

\item{slices}{Optional. Integer vector of slice positions in the selected \code{plane}
to keep. If \code{NULL}, all slices are retained.}

\item{lim}{Defines the 2D spatial limits of the resulting \code{vbl2D}. By default, the
\link[=ccs_limits]{coordinate-space} limits of the input vibble are and mapped to
col and rowdepending on \code{plane}. User can adjust that by supplying:

\itemize{
\item  a valid \link[=is_limit]{limit}: a numeric vector of length two which is recycled for col and row.
\item a list with \code{col} and \code{row} elements, each being a \link[=is_limit]{limit} for the respective axis.
}

When the supplied \link[=is_bb2D]{2D bounding-box} is smaller than the slice extents,
only voxels inside this region are retained, giving \code{lim} a filtering effect.}

\item{expand}{An \link[=is_expand]{expand} specification applied to the col/row limits \strong{after}
the interpretation and potential filtering of \code{lim}.}

\item{.cond}{Optional. An expression evaluated with \link[rlang:args_data_masking]{data-masking}
to filter voxels. If \code{NULL}, all voxels are retained.}

\item{.by}{Optional. A \link[dplyr:dplyr_tidy_select]{tidy-selection} of columns to
group by before applying the filtering logic of \code{.cond}.}
}
\value{
A \code{tibble} with class \code{vbl2D} containing columns \code{col}, \code{row}, \code{slice},
and all remaining variables from \code{vbl}. Attributes store the used plane and any
applied offsets.
}
\description{
Convert a 3D \code{vbl} object into a 2D representation for a given anatomical plane and selected slices. Optionally apply filtering conditions and spatial offsets for visualization.
}
\details{
The function derives the required axes using \link{req_axes_2D}() and
renames the corresponding coordinate columns to \code{col}, \code{row}, and \code{slice},
and optionally filters by \code{slices}, \code{lim} and \code{.cond}.

\itemize{
\item \code{col} and \code{row} store the 2D coordinates in voxel space.
\item \code{slice} stores the position along the third axis in the selected \code{plane}.
\item Attributes \code{"offset_dist"}, \code{"offset_dir"}, and \code{"plane"}
describe the applied spatial offset and the original plane.
}
}
