% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibble2D.R
\name{vibble2D}
\alias{vibble2D}
\title{Create a 2D vibble from 3D voxel data}
\usage{
vibble2D(
  vbl,
  plane,
  slices = NULL,
  crop = NULL,
  expand = FALSE,
  .cond = NULL,
  .by = NULL
)
}
\arguments{
\item{vbl}{A \link[=vbl_doc_vbl]{vibble}.}

\item{plane}{Character scalar. The anatomical orientation, one of \emph{'sag', 'axi' or 'cor'}.
Defaults to the \link[=vbl_doc_defaults]{global default plane}.}

\item{slices}{Optional. Integer vector of slice positions in the selected \code{plane}
to keep. If \code{NULL}, all slices are retained.}

\item{crop}{Defines the \link[=vbl_doc_ref_bb]{data bounding box}
applied to all slices in native (pre-offset) coordinates.

\itemize{
\item{\code{NULL}: }{ No cropping.}
\item{\code{limit}: }{ A single spatial \link[=is_limit]{limit} recycled for both \code{col} and \code{row}.}
\item{\code{bb2D}: }{ A \link[=is_bb2D]{2D bounding box} with separate limits for \code{col} and \code{row}.}
}

When the supplied bounding box is smaller than the slice extents, only voxels
inside this region are retained, giving \code{crop} a filtering effect.}

\item{expand}{An \link[=is_expand]{expand} specification applied to the col/row limits
of \code{crop} \strong{after} interpretation and potential filtering to set the \link[=vbl_doc_ref_bb]{screen bounding box}.}

\item{.cond}{An expression evaluated with \link[rlang:args_data_masking]{data-masking}
to filter voxels. If \code{NULL}, all voxels are retained.}

\item{.by}{A \link[dplyr:dplyr_tidy_select]{tidy-selection} of columns to
group by before applying the filtering logic of \code{.cond}.}
}
\value{
A \link[=vbl_doc_vbl2D]{2D vibble}.
}
\description{
Convert a 3D \code{vbl} object into a 2D representation for a given
anatomical plane and selected slices. Optionally apply filtering conditions.
}
